+++ test/cubit/project_detail_cubit_test.dart
@@ -10,7 +10,7 @@ import 'package:deadbolt/models/timelock_types.dart';
 import 'package:deadbolt/services/project_descriptor_service.dart';
 import 'package:deadbolt/src/rust/api/model.dart';
 
-import '../test_utils.dart';
+import '../fixtures/sqlite_setup.dart';
 
 class MockProjectDescriptorService extends Mock
     implements ProjectDescriptorService {}